Do not proxy from CLI container
It seems to cause too many problems at build time. Instead we run
cron through FPM. As the scope of the DrupalModules context increases
we rename it to DrupalContext.
